fix: refactor hint
