--- src/components/form/field-hint/index.ts
@@ -1,2 +1,0 @@
-export * from './field-hint';
-export * from './types';
