--- src/components/form/field-hint/types.ts
@@ -1,11 +1,0 @@
-export type FieldHintProps = {
-    hintName: string;
-    hintText: string | null | undefined;
-    errorText: string | null | undefined;
-    status: TextStatusVariant;
-};
-
-export enum TextStatusVariant {
-    HINT = 'hint',
-    ERROR = 'error',
-}
