+++ src/components/form/hint/hint.tsx
@@ -1,0 +1,16 @@
+import { FC } from 'react';
+import { HintProps, HintVariant } from './types';
+
+export const Hint: FC<HintProps> = ({ id, children, status, ...props }) => {
+    const isError = status === HintVariant.ERROR;
+    return (
+        <span
+            id={`${id}-${status}`}
+            className={isError ? 'self-end text-error' : 'text-secondary-400'}
+            role={isError ? 'alert' : undefined}
+            {...props}
+        >
+            {children}
+        </span>
+    );
+};
