+++ src/components/form/hint/index.ts
@@ -1,0 +1,2 @@
+export * from './hint';
+export * from './types';
