+++ src/components/form/index.ts
@@ -1,4 +1,4 @@
 export * from './input';
 export * from './textarea';
 export * from './file-upload';
-export * from './field-hint';
+export * from './hint';
