+++ src/components/form/input/input.tsx
@@ -2,7 +2,7 @@ import { FC, useState } from 'react';
 import { InputProps, InputType } from './types';
 import { Label, LabelSize } from '../../typography';
 import { IconSize } from '../../../elements';
-import { FieldHint } from '../field-hint';
+import { Hint } from '../hint';
 import clsx from 'clsx';
 
 export const Input: FC<InputProps> = ({
@@ -23,7 +23,7 @@ export const Input: FC<InputProps> = ({
     ...props
 }) => {
     const [isPrivacy, setIsPrivacy] = useState(false);
-    const inputWrapperClasses = 'relative w-full h-fit wrap-input';
+    const inputWrapperClasses = 'relative flex flex-col w-full h-fit wrap-input';
     const inputClasses = clsx(
         'pr-10 h-xl w-full rounded-s px-s py-s text-secondary-700 placeholder-secondary-300 focus:outline-2 focus:outline-primary-600',
         className,
@@ -74,12 +74,7 @@ export const Input: FC<InputProps> = ({
                 )}
             </div>
             {hintName && status && (hintText || errorText) && (
-                <FieldHint
-                    status={status}
-                    hintName={hintName}
-                    errorText={errorText}
-                    hintText={hintText}
-                />
+                <Hint status={status}>{errorText || hintText}</Hint>
             )}
         </div>
     );
